ajustes

+++ booking-service/src/main/java/com/orbitbook/booking/dto/destination/DestinationResponseDTO.java
@@ -1,11 +1,8 @@
 package com.orbitbook.booking.dto.destination;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -23,7 +20,7 @@ public class DestinationResponseDTO {
 
     private Integer distanceKm;
 
-    private Integer basePrice;
+    private BigDecimal basePrice;
 
     private Integer capacity;
