+++ booking-service/src/main/java/com/orbitbook/booking/dto/destination/DestinationUpdateDTO.java
@@ -1,10 +1,8 @@
 package com.orbitbook.booking.dto.destination;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -19,7 +17,7 @@ public class DestinationUpdateDTO {
 
     private Integer distanceKm;
 
-    private Integer basePrice;
+    private BigDecimal basePrice;
 
     private Integer capacity;
